refactor(core): change directory for emitter
